Move KnativeServing test functions to v1alpha1 package
